+++ CHANGELOG.md
@@ -53,6 +53,9 @@
   - `Belt.MutableMap.getExn` → `Belt.MutableMap.getOrThrow`
   - `Belt.Set.getExn` → `Belt.Set.getOrThrow`
   - `Belt.MutableSet.getExn` → `Belt.MutableSet.getOrThrow`
+  - `Belt.List.getExn` → `Belt.List.getOrThrow`
+  - `Belt.List.tailExn` → `Belt.List.tailOrThrow`
+  - `Belt.List.headExn` → `Belt.List.headOrThrow`
   - Old functions remain available but are marked as deprecated with guidance to use the new `OrThrow` variants.
   - https://github.com/rescript-lang/rescript/pull/7518, https://github.com/rescript-lang/rescript/pull/7554, https://github.com/rescript-lang/rescript/pull/7581